Complete "required" Theme UI colors
`accent` and `muted` aren't actually used by any Theme UI components, so
they'll only be used when manually specified, which you won't need to
with Bootstrap's existing colors. `$text-muted` is a Bootstrap 5
variable, but it isn't a main one, and `textMuted` isn't being used in
any existing components or styles in this repo, so I'm deleting it for
now.